Fixing weather-mv refactor.
- Moved a lot of BigQuery related concerns out of the main pipeline.
- Updated -i help comment
- Added arg validation to Sink interface.
- Fixed a lot of errors where BQ logic was applied to regridding.
